requirements.txt
# Core Framework
Django==5.0.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2

# CORS & Security
django-cors-headers==4.3.1

# Database
# Using SQLite3 for local development (included with Python)

# Environment & Config
python-dotenv==1.0.1

# HTTP Requests (for seeding data from APIs)
requests==2.31.0

# Production Server
gunicorn==22.0.0

# Testing & Development
pytest==8.3.4
pytest-django==4.9.0
coverage==7.6.9

# Code Quality
ruff==0.8.4
mypy==1.13.0
django-stubs==5.1.1

